:truck: Move actions and reducers to state folder
